refactor: client index-exports to exports.ts
