+++ packages/client/src/index.ts
@@ -7,6 +7,6 @@
 
 import { StreamrClient } from './StreamrClient'
 
-export * from './index-exports'
+export * from './exports'
 export default StreamrClient
 // Note awful export wrappers in index-commonjs.js & index-esm.mjs
